+++ src/components/Views/Calendar/Calendar.js
@@ -1,0 +1,22 @@
+import React from 'react';
+
+export default class Calendar extends React.Component {
+    constructor(props) {
+        super(props);
+        this.state = {
+
+        }
+    }
+
+    componentDidMount() {
+        console.log("test cal");
+    }
+
+    render() {
+        return(
+            <div>
+                <p>test</p>
+            </div>
+        )
+    }
+}
